requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap4==2022.1
django-filter==23.4
django-extensions==3.2.3
geopy==2.4.0
requests==2.31.0